feat: Add multi-recipient transfer example
Adds an example of a multi-recipient transfer using the `transfer` method in the TariWalletGrpcClient. This new feature allows sending funds to multiple recipients in a single transaction, which can be enabled by setting the `singleTx` parameter to `true` within the transfer request. The example includes defining multiple recipient objects with different addresses, amounts, feePerGram and payment types.